PlaneChange method + little changes with ArrayList<>

+++ src/Aircraft/Plane.java
@@ -40,7 +40,7 @@ public abstract class Plane {
         this.flyDistance = flyDistance;
     }
 
-    public void setFuelConsumption(int fuelConsumption) {
+    public void setFuelConsumption(double fuelConsumption) {
         this.fuelConsumption = fuelConsumption;
     }
 }
